Utført: sjekker nå om bruker finnes allerede før man setter den inn . Sjekk av postback og sette rett valgt bruker på kontroller om postback er kjørt. Gui holder rette deler av menyen i orden ved postback i kontrollpanelet. Har lagt til validationgroups i AdminArea. Slett bruker knapp knapp aktivert bare hvis man er Admin.

+++ Blodbanken/Controls/UserDeleteControl.ascx.cs
@@ -11,10 +11,17 @@ namespace Blodbanken.Controls {
       public string CurrentUser { get; set; }
       AuthenticatonModule AuthMod = new AuthenticatonModule();
       protected void Page_Load(object sender, EventArgs e) {
+         System.Security.Principal.GenericPrincipal myUser = (System.Security.Principal.GenericPrincipal)HttpContext.Current.Cache.Get("customPrincipal");
+         if (myUser != null)
+            HttpContext.Current.User = myUser;
 
+         if (HttpContext.Current.User.IsInRole("Admin")) btnDeleteUser.Visible = true;
+         else btnDeleteUser.Visible = false;
       }
-      public void DeleteUser(object sender, EventArgs e) {
-         AuthMod.DeleteUser(CurrentUser);
+      public void DeleteUser(object sender, CommandEventArgs e) {
+         if (e.CommandName == btnDeleteUser.CommandName) {
+            AuthMod.DeleteUser(CurrentUser);
+         }         
       }
    }
 }
